+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 * Fixed `ffmpeg_animate_frames` to handle paths and ndarrays
 
+### Added
+* Add Palette class
+
 
 ## Version 0.5.2 - Released 2024-09-09